Is Matchbox Dead?
This project seems intentionally hostile to use, with no issues available, and precious little documentation.  The gRPC example no longer appears to work either, and the only way of getting in touch appears to be to open a pull request, which I'm doing here.

So far I've tried creating the certs and starting up everything pointing at the right files, and bootcmd just hangs forever.  When using the terraform provider it does eventually time out due to the context expiring, but it also doesn't do anything useful either.

If the Poseidon Labs/matchbox stack is not actually intended to be used that's fine, I can go find some other software, but this does feel like a bug somewhere in the matchbox code.